Move Processing into its own state
Smaller state is better less re-renders
