+++ Source/TimeWarp.Architecture.Template/templates/TimeWarp.Architecture/Source/ContainerApps/Web/Web.Spa/Features/Processing/Actions/ProcessingState.TwoSecondTask.cs
@@ -1,6 +1,6 @@
-namespace TimeWarp.Architecture.Features.Applications;
+namespace TimeWarp.Architecture.Features.Processing;
 
-internal partial class ApplicationState
+internal partial class ProcessingState
 {
   public static class TwoSecondTask
   {
