+++ Source/TimeWarp.Architecture.Template/templates/TimeWarp.Architecture/Source/ContainerApps/Web/Web.Spa/Features/Processing/Pipeline/TrackProcessingAttribute.cs
@@ -1,4 +1,4 @@
-﻿namespace TimeWarp.Architecture.Features.Applications;
+﻿namespace TimeWarp.Architecture.Features.Processing;
 
 [AttributeUsage(AttributeTargets.Class, AllowMultiple = false)]
 public class TrackProcessingAttribute : Attribute { }
